Create EarlyStopping class. Rename file with train function.
